# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vocal Mind API",
    "description": "Backend API for Vocal Mind - RAG, ASR, and more.",
    "version": "0.1.0"
  },
  "paths": {
    "/rag/health": {
      "get": {
        "tags": [
          "RAG"
        ],
        "summary": "Health Check",
        "operationId": "health_check_rag_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rag/query": {
      "post": {
        "tags": [
          "RAG"
        ],
        "summary": "Query",
        "operationId": "query_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rag/ingest": {
      "post": {
        "tags": [
          "RAG"
        ],
        "summary": "Ingest",
        "operationId": "ingest_rag_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Global health check.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "components": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Components"
          }
        },
        "type": "object",
        "required": [
          "status",
          "components"
        ],
        "title": "HealthResponse"
      },
      "IngestRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "IngestRequest"
      },
      "IngestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "documents_processed": {
            "type": "integer",
            "title": "Documents Processed"
          },
          "nodes_created": {
            "type": "integer",
            "title": "Nodes Created"
          }
        },
        "type": "object",
        "required": [
          "status",
          "documents_processed",
          "nodes_created"
        ],
        "title": "IngestResponse"
      },
      "QueryRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Question to ask"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceNode"
            },
            "type": "array",
            "title": "Sources"
          },
          "timing": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Timing"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "timing"
        ],
        "title": "QueryResponse"
      },
      "SourceNode": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "score",
          "text",
          "metadata"
        ],
        "title": "SourceNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}